Improve and Fix The Sequence Editor
The sequence editor's v-for was using an idex as the :key value.
This causes vuedraggable to do weird stuff. This change creates a
unified loading/saving utility for automations. It injects a fast
nanoid into every action that can be keyed off of.

This change also prevents the header from disabling itself completely.
v-if on the slot for vuedraggable causes issues. Instead we use v-if on
a card inside of a slot div.

Mark sequence items as is-draggable so header and footer divs don't get
moved around.