Fix out of order selection bug in sequence editor.
vue-draggable appears to not properly update the virtual dom ordering
of sequence items. This causes selection of newly added actions to go
all wonky like. To fix this we make sure to explicitly derive indices
from action keys instead of assuming the $ref array is in the proper
order. We can throw this fix in the garbage when we move to timelines.