Add websocket action
This commit adds a general plugin for interacting with programs that provide an API via websockets.

It only supports a singular software to be used for simplicity. The user is provided an action that can be used to send an arbitrary message via the configured Websocket.

Amongst others, this allows interaction with VTuber Plus.