The Great Refactorication Pt 2: Electric Boogaloo
Switched to using yarn 3 from pnpm. Electron-builder does not support pnpm

After MANY MANY MANY attempts. I've settled for running two concurrent
vite dev servers. One for CastMate electron and one for CastMate overlays.

multi-vite.mjs is a utility to launch all the instances of vite for building
and serving, out of the correct working directories since yarn's workspace
system does not do this correctly.

In a functioning world, we would only need to bundle the overlay components
once. However, this is javascript and apparently in the whole history
of the language no one wanted to share components between two pages.

Instead we suffer the shame of bundling our overlay components twice.

TODO: Forward internal server for overlays to the dev server when not
in packaged builds.
TODO: Figure out electron-builder call so overlay files are available
and served.