Mega Commit: State Changes, State in overlays
Sorry this commit is huge, (not really sorry).

This change adds a new StateManager which is a singleton for managing the reactivity of plugin state, forwarding that state to the UI, and constructing template contexts.

Setup RemoteTemplateManager which allows the UI to evaluate templates on the main process. This way templateFunctions can be used. This unfortunately needed the creation of a remoteTemplates vuex store. We'll need to convert this to pinia at some point.

Made main process reactivity async compatible using async hooks.

Reorganized a few files

Cleaned up some log statements

Organized Plugin a little bit