Fix Double Init and Missing Elgato Lights
Vite was improperly emitting the imports for castmate-overlay-components.

This bugged import was causing the electron browser to load main.js from two slightly different sources.

This caused the main processes' stores to be inited twice.

This caused any IPC functions to be registered twice.

This is likely the source of the double sound plays.

This also caused resource inits to be wonky causing Elgato lights to occasionally not show up in the resource list.

* Upgraded vite to 4.x
* Made sure to pass the "up" function to bonjour on construction so it's not a data race to assign a handler.