Upgrade to Vite 4, Fix the double import bug.
castmate-overlay-components was configured in such a bizarre way it was causing main.js to load twice.

This is probably the root cause of the double sounds.

It's definitely what caused the elgato lights to quantumly dissappear.

Now both castmate and the overlay properly use the component package.

The component package is built with vite, and then the built version is bundled.

In serve we still alias to HMR.