First pass at loading plugins
Eventually this should be dynamically loaded, but vite and node.js don't
want us to acheive our dreams. Instead we just explicitly load them.

In the future we want to use something like pluggable-electron to load
them. Which would work if we used vite to build each plugin separately.

However, we want HMR for UI and to reload the whole app when the main
process changes so we don't slowly lose our minds building this thing.

So instead we're stuck with manually adding each plugin as a dep and
manually loading it in. For now.

We'll need to talk to some sort of vite expert to get this to work.