Make Action Palette Work, Fix Drop Zones, Support floating Sequences
Floating Sequences are sequences that aren't attached to the trigger root

When an action palette item is selected, it's spawned in as a new floating sequence

This lets us finally test our drag and drop.

Because moved actions are created before their former location is unmounted
we have to store an array of drop zones in the root.

We had to add keys to the dropzones so they don't get reused by the VDOM
ruining our dropKeys / lookup table