Track "Local" drops in automations
Since "dropped" events fires before "dragend" events, if an action
was moved earlier in the chain than it's current position. The v-model
changes would cause the original VDOM element to destroy itself
the dragend event would then not fire.

So if a drag is made "locally" in that it's being moved in the same
automation edit, we do the removal before the placement in dropped.
Then set the dataTransfer action to none.