Add a Duration Input
Using google's timer widget as a guide this change creates a better time input.

The duration input works by having a hidden <input> tag and then rendering
the text, cursor, and selections itself so that the spacing characters
that indicate hours, minutes, and seconds aren't part of the actual input

This removes any of the weirdness caused by normal masking.
As a consequence the selections look super weird bro.