Potentially Fix Twitch Ad Issues
Handle case where the user's system clock is more than 1 second ahead of
the twitch server clock.

I think what was happening is the ad schedule validation query happened
AFTER the theoretical start time on the server. Which meant that it was
updating the schedule for the *next* ad.

So this change makes the validation query meant to catch snoozes
only reschedules if the snooze count has gone down.